feat(crc): it is possible to pass the checksum function in the CRC field parameters
If you pass the checksum function as parameters, then your structure will have the static `crc`
method that you can use to calculate and update this field. This changed the behavior of the static
`crc` method.

BREAKING CHANGE: The static `crc` method has been changed. It can now be used to calculate the
checksum if you have specified a checksum function for this field.